Move more limits into limits file.
