+++ src/tools/peptide-search/components/PeptideSearchForm.tsx
@@ -28,6 +28,11 @@ import {
 
 import { truncateTaxonLabel } from '../../utils';
 
+import {
+  PEPTIDE_SEARCH_SEQUENCES_COUNT,
+  PEPTIDE_SEARCH_SEQ_MINIMUM_LENGTH,
+} from '../../../shared/config/limits';
+
 import { JobTypes } from '../../types/toolsJobTypes';
 import { FormParameters } from '../types/peptideSearchFormParameters';
 import { peps, lEQi, spOnly } from '../types/peptideSearchServerParameters';
@@ -48,10 +53,6 @@ import {
 
 import sticky from '../../../shared/styles/sticky.module.scss';
 import '../../styles/ToolsForm.scss';
-
-// just because, no actual known limit
-export const PEPTIDE_SEARCH_SEQ_MINIMUM_LENGTH = 7;
-export const PEPTIDE_SEARCH_SEQUENCES_COUNT = 100;
 
 const title = namespaceAndToolsLabels[JobTypes.PEPTIDE_SEARCH];
