+++ src/tools/peptide-search/state/peptideSearchFormReducer.ts
@@ -1,7 +1,7 @@
 import { ActionType } from 'typesafe-actions';
 import { truncate } from 'lodash-es';
 
-import { PEPTIDE_SEARCH_SEQUENCES_COUNT } from '../components/PeptideSearchForm';
+import { PEPTIDE_SEARCH_SEQUENCES_COUNT } from '../../../shared/config/limits';
 import * as peptideSearchFormActions from './peptideSearchFormActions';
 import splitAndTidyText from '../../../shared/utils/splitAndTidyText';
